`rm` safety checks

+++ data/common-userfiles_3-clean-root.sh
@@ -7,6 +7,8 @@ cd "${DIRECTORY}" || exit 1
 # removing common files after building, just for tidiness
 for i in $(find . -mindepth 1 -printf '%f\n'); do
 	for f in etc/skel root home/hacker; do
-		rm -rf "../../${FILE_OVERLAY}/${f}/${i}"
+		if [ -e "${i}" ]; then
+			rm -rf "../../${FILE_OVERLAY}/${f}/${i}"
+		fi
 	done
 done
